old_version/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pandas==1.3.3
pyyaml==5.4.1
watchdog==2.1.5
boto3==1.18.50
requests==2.26.0 